LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-svg-script-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

Harness Error (TIMEOUT), message = null

PASS Assign String to SVGScriptElement.innerHTML.
PASS Assign TrustedHTML to SVGScriptElement.innerHTML.
TIMEOUT Assign TrustedHTML to SVGScriptElement.innerHTML and execute it. Test timed out
NOTRUN Modify SVGScriptElement via DOM manipulation.
NOTRUN Assign string to SVGScriptElement.href.baseVal.
NOTRUN Assign TrustedScriptURL to SVGScriptElement.href.baseVal.
NOTRUN Assign string to non-attached SVGScriptElement.href via setAttribute.
NOTRUN Assign TrustedScriptURL to non-attached SVGScriptElement.href via setAttribute.
NOTRUN Assign string to attached SVGScriptElement.href via setAttribute.
NOTRUN Assign TrustedScriptURL to attached SVGScriptElement.href via setAttribute.
NOTRUN Setup default policy
NOTRUN Assign String to SVGScriptElement.innerHTML w/ default policy.
NOTRUN Assign string to SVGScriptElement.href.baseVal  w/ default policy.
NOTRUN Assign string to non-attached SVGScriptElement.href via setAttribute w/ default policy.
NOTRUN Assign string to attached SVGScriptElement.href via setAttribute w/ default policy.
